P3: year rev. 1

+++ p3/IndiceSimple.java
@@ -72,7 +72,7 @@ public class indiceSimple {
                 doc.add(new TextField("title", r[2], Field.Store.YES));
                 if(!r[3].equals("")){
                     doc.add(new SortedNumericDocValuesField("year", Long.parseLong(r[3])));
-                    doc.add(new StoredField("year", Integer.parseInt(r[3])));
+                    doc.add(new StoredField("year", Long.parseLong(r[3])));
                 }
                 doc.add(new TextField("source_title", r[4], Field.Store.YES));
                 doc.add(new StoredField("volume", r[5]));
